adjustments, move everything
