+++ packages/dbos-openapi/Gruntfile.js
@@ -1,0 +1,9 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+var nbgv = require('nerdbank-gitversioning')
+
+module.exports = function (grunt) {
+    grunt.registerTask('setversion', function () {
+        var done = this.async();
+        nbgv.setPackageVersion().then(() => done());
+    });
+};
