+++ packages/dbos-openapi/tsDiagUtil.ts
