+++ packages/dbos-openapi/tsconfig.json
@@ -1,0 +1,30 @@
+/* Visit https://aka.ms/tsconfig to read more about this file */
+{
+  "compilerOptions": {
+    "target": "esnext",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
+    "experimentalDecorators": true,                      /* Enable experimental support for legacy experimental decorators. */
+    "emitDecoratorMetadata": true,                       /* Emit design-type metadata for decorated declarations in source files. */
+    "module": "commonjs",                                /* Specify what module code is generated. */
+    "declaration": true,                                 /* Generate .d.ts files from TypeScript and JavaScript files in your project. */
+    "declarationMap": true,                              /* Create sourcemaps for d.ts files. */
+    "sourceMap": true,                                   /* Create source map files for emitted JavaScript files. */
+    "outDir": "./dist",                                  /* Specify an output folder for all emitted files. */
+    "newLine": "lf",                                     /* Set the newline character for emitting files. */
+    "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
+    "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
+    "strict": true,                                      /* Enable all strict type-checking options. */
+    "skipLibCheck": true,                                /* Skip type checking all .d.ts files. */
+    "paths": {
+      "../../src/telemetry/*": ["../../src/telemetry/*"],
+      "../../src/utils": ["../../src/utils"],
+      "../../src/user_database": ["../../src/user_database"],
+      "../../src/system_database": ["../../src/system_database"],
+      "../../src/dbos-runtime/*": ["../../src/dbos-runtime/*"]
+    }
+  },
+  "include": [                                         /* Specifies an array of filenames or patterns to include in the program. */
+    ".",
+    "applications"
+  ],
+ /* "references": [{"path": "../.."}]*/
+}
